requirements.txt
python_telegram_bot==12.3.0
python_dateutil==2.8.1
